+++ web/src/typing_events.ts
@@ -3,6 +3,7 @@ import assert from "minimalistic-assert";
 
 import render_typing_notifications from "../templates/typing_notifications.hbs";
 
+import * as message_lists from "./message_lists";
 import * as narrow_state from "./narrow_state";
 import * as people from "./people";
 import {current_user, realm} from "./state_data";
@@ -45,6 +46,13 @@ type TypingEvent = {
           sender: UserInfo;
       }
 );
+
+type TypingMessageEditEvent = {
+    message_id: number;
+    op: "start" | "stop";
+    type: "typing_edit_message";
+    sender: UserInfo;
+};
 
 function get_users_typing_for_narrow(): number[] {
     if (narrow_state.narrowed_by_topic_reply()) {
@@ -107,6 +115,24 @@ export function render_notifications_for_narrow(): void {
     }
 }
 
+function apply_message_edit_notifications($row: JQuery, is_typing: boolean): void {
+    const $editing_notifications = $row.find("#editing_notifications");
+    if (is_typing) {
+        $row.find(".message_edit_notice").hide();
+        $editing_notifications.show();
+    } else {
+        $row.find(".message_edit_notice").show();
+        $editing_notifications.hide();
+    }
+}
+
+function render_message_editing_typing(message_id: number, is_typing: boolean): void {
+    const $row = message_lists.current?.get_row(message_id);
+    if ($row !== undefined) {
+        apply_message_edit_notifications($row, is_typing);
+    }
+}
+
 function get_key(event: TypingEvent): string {
     if (event.message_type === "stream") {
         return typing_data.get_topic_key(event.stream_id, event.topic);
@@ -130,6 +156,16 @@ export function hide_notification(event: TypingEvent): void {
     }
 }
 
+export function hide_message_edit_notification(event: TypingMessageEditEvent): void {
+    const message_id = event.message_id;
+    const key = JSON.stringify(message_id);
+    typing_data.clear_inbound_timer(key);
+    const removed = typing_data.remove_edit_message_typing_id(message_id);
+    if (removed) {
+        render_message_editing_typing(message_id, false);
+    }
+}
+
 export function display_notification(event: TypingEvent): void {
     const sender_id = event.sender.user_id;
 
@@ -143,6 +179,20 @@ export function display_notification(event: TypingEvent): void {
         realm.server_typing_started_expiry_period_milliseconds,
         () => {
             hide_notification(event);
+        },
+    );
+}
+
+export function display_message_edit_notification(event: TypingMessageEditEvent): void {
+    const message_id = event.message_id;
+    const key = JSON.stringify(message_id);
+    typing_data.add_edit_message_typing_id(message_id);
+    render_message_editing_typing(message_id, true);
+    typing_data.kickstart_inbound_timer(
+        key,
+        realm.server_typing_started_expiry_period_milliseconds,
+        () => {
+            hide_message_edit_notification(event);
         },
     );
 }